+++ controller/konnect/ops/zz_generated_ops_getforuid.go
@@ -102,6 +102,8 @@ func getForUID[
 		return getPortalEmailConfigForUID(ctx, sdk.GetPortalEmailsSDK(), ent)
 	case *konnectv1alpha1.PortalIPAllowList:
 		return getPortalIPAllowListForUID(ctx, sdk.GetPortalsIPAllowListSDK(), ent)
+	case *konnectv1alpha1.PortalIdentityProviderRequest:
+		return getPortalIdentityProviderRequestForUID(ctx, sdk.GetPortalAuthSettingsSDK(), ent)
 	case *konnectv1alpha1.PortalPage:
 		return getPortalPageForUID(ctx, sdk.GetPortalPagesSDK(), ent)
 	case *konnectv1alpha1.PortalTeam: